+++ ts/features/bonus/bpd/saga/networking/index.ts
@@ -9,47 +9,6 @@ import { SagaCallReturnType } from "../../../../../types/utils";
 import { BackendBpdClient } from "../../api/backendBpdClient";
 import { bpdLoadActivationStatus } from "../../store/actions/details";
 
-export function* executeAndDispatch(
-  remoteCall:
-    | ReturnType<typeof BackendBpdClient>["enrollCitizenIO"]
-    | ReturnType<typeof BackendBpdClient>["find"],
-  action: typeof bpdEnrollUserToProgram | typeof bpdLoadActivationStatus
-) {
-  try {
-    const enrollCitizenIOResult: SagaCallReturnType<typeof remoteCall> = yield call(
-      remoteCall,
-      // due to avoid required headers coming from code autogenerate
-      // (note the required header will be injected automatically)
-      {} as any
-    );
-    if (enrollCitizenIOResult.isRight()) {
-      if (enrollCitizenIOResult.value.status === 200) {
-        const { enabled, payoffInstr } = enrollCitizenIOResult.value.value;
-        yield put(
-          action.success({
-            enabled,
-            payoffInstr
-          })
-        );
-        return;
-      } else if (enrollCitizenIOResult.value.status === 404) {
-        yield put(
-          action.success({
-            enabled: false,
-            payoffInstr: undefined
-          })
-        );
-        return;
-      }
-      throw new Error(`response status ${enrollCitizenIOResult.value.status}`);
-    } else {
-      throw new Error(readableReport(enrollCitizenIOResult.value));
-    }
-  } catch (e) {
-    yield put(action.failure(e));
-  }
-}
-
 export function* executeAndDispatchV2(
   remoteCall:
     | ReturnType<typeof BackendBpdClient>["enrollCitizenV2IO"]
@@ -97,30 +56,10 @@ export function* executeAndDispatchV2(
   }
 }
 
-/**
- * make a request to get the citizen status
- * @deprecated
- */
-export function* getCitizen(
-  findCitizen: ReturnType<typeof BackendBpdClient>["find"]
-): SagaIterator {
-  yield call(executeAndDispatch, findCitizen, bpdLoadActivationStatus);
-}
-
 export function* getCitizenV2(
   findCitizen: ReturnType<typeof BackendBpdClient>["findV2"]
 ): SagaIterator {
   yield call(executeAndDispatchV2, findCitizen, bpdLoadActivationStatus);
-}
-
-/**
- * make a request to enroll the citizen to the bpd
- * @deprecated
- */
-export function* putEnrollCitizen(
-  enrollCitizenIO: ReturnType<typeof BackendBpdClient>["enrollCitizenIO"]
-): SagaIterator {
-  yield call(executeAndDispatch, enrollCitizenIO, bpdEnrollUserToProgram);
 }
 
 export function* putEnrollCitizenV2(