+++ ts/features/bonus/siciliaVola/store/actions/index.ts
@@ -1,0 +1,3 @@
+import { SvVoucherGenerationActions } from "./voucherGeneration";
+
+export type SvActions = SvVoucherGenerationActions;