+++ ts/features/euCovidCert/screens/BaseEuCovidCertificateLayout.tsx
@@ -28,7 +28,7 @@ const styles = StyleSheet.create({
   }
 });
 
-const Header = () => (
+export const Header = () => (
   <>
     <View style={styles.row}>
       <H1 style={IOStyles.flex}>
@@ -48,7 +48,11 @@ const Header = () => (
 export const BaseEuCovidCertificateLayout = (props: Props) => {
   const elementRef = useRef(null);
   return (
-    <BaseScreenComponent goBack={true} contextualHelp={emptyContextualHelp}>
+    <BaseScreenComponent
+      goBack={true}
+      contextualHelp={emptyContextualHelp}
+      shouldAskForScreenshotWithInitialValue={false}
+    >
       <NavigationEvents onDidFocus={() => setAccessibilityFocus(elementRef)} />
       <SafeAreaView
         style={IOStyles.flex}