+++ ts/features/euCovidCert/screens/valid/EuCovidCertValidScreen.tsx
@@ -1,11 +1,20 @@
-import { constNull } from "fp-ts/lib/function";
 import { View } from "native-base";
 import * as React from "react";
-import { Dimensions, Image, StyleSheet, TouchableOpacity } from "react-native";
+import { useState } from "react";
+import {
+  Dimensions,
+  Image,
+  StyleProp,
+  StyleSheet,
+  TouchableOpacity,
+  ViewStyle
+} from "react-native";
+import { CaptureOptions } from "react-native-view-shot";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
 import FooterWithButtons from "../../../../components/ui/FooterWithButtons";
 import I18n from "../../../../i18n";
+import { mixpanelTrack } from "../../../../mixpanel";
 import { GlobalState } from "../../../../store/reducers/types";
 import themeVariables from "../../../../theme/variables";
 import {
@@ -18,7 +27,24 @@ import {
   navigateToEuCovidCertificateQrCodeFullScreen
 } from "../../navigation/actions";
 import { ValidCertificate } from "../../types/EUCovidCertificate";
-import { BaseEuCovidCertificateLayout } from "../BaseEuCovidCertificateLayout";
+import {
+  BaseEuCovidCertificateLayout,
+  Header
+} from "../BaseEuCovidCertificateLayout";
+import { useIOBottomSheet } from "../../../../utils/bottomSheet";
+import ButtonDefaultOpacity from "../../../../components/ButtonDefaultOpacity";
+import { IOStyles } from "../../../../components/core/variables/IOStyles";
+import { H3 } from "../../../../components/core/typography/H3";
+import { H5 } from "../../../../components/core/typography/H5";
+import IconFont from "../../../../components/ui/IconFont";
+import { IOColors } from "../../../../components/core/variables/IOColors";
+import { showToast } from "../../../../utils/showToast";
+import { captureScreenShoot } from "../../utils/screenshoot";
+import {
+  FlashAnimatedComponent,
+  FlashAnimationState
+} from "../../components/FlashAnimatedComponent";
+import { euCovidCertCurrentSelector } from "../../store/reducers/current";
 
 type OwnProps = {
   validCertificate: ValidCertificate;
@@ -30,6 +56,22 @@ const styles = StyleSheet.create({
     width: Dimensions.get("window").width - themeVariables.contentPadding * 2,
     height: Dimensions.get("window").width - themeVariables.contentPadding * 2,
     flex: 1
+  },
+  container: {
+    paddingRight: 0,
+    paddingLeft: 0,
+    marginVertical: 20,
+    height: 60,
+    backgroundColor: IOColors.white
+  },
+  flexColumn: {
+    flexDirection: "column",
+    flex: 1
+  },
+  row: {
+    flexDirection: "row",
+    alignItems: "flex-start",
+    justifyContent: "space-between"
   }
 });
 
@@ -37,48 +79,124 @@ type Props = ReturnType<typeof mapDispatchToProps> &
   ReturnType<typeof mapStateToProps> &
   OwnProps;
 
-const EuCovidCertValidComponent = (props: Props): React.ReactElement => (
+type EuCovidCertValidComponentProps = Props & {
+  markdownWebViewStyle?: StyleProp<ViewStyle>;
+  messageId?: string;
+};
+const EuCovidCertValidComponent = (
+  props: EuCovidCertValidComponentProps
+): React.ReactElement => (
   <View>
     {props.validCertificate.qrCode.mimeType === "image/png" && (
-      <TouchableOpacity
-        testID={"QRCode"}
-        accessible={true}
-        accessibilityRole={"imagebutton"}
-        accessibilityLabel={I18n.t(
-          "features.euCovidCertificate.valid.accessibility.qrCode"
-        )}
-        accessibilityHint={I18n.t(
-          "features.euCovidCertificate.valid.accessibility.hint"
-        )}
-        onPress={() =>
-          props.navigateToQrCodeFullScreen(
-            props.validCertificate.qrCode.content
-          )
-        }
-      >
-        <Image
-          source={{
-            uri: `data:image/png;base64,${props.validCertificate.qrCode.content}`
-          }}
-          style={styles.qrCode}
-        />
-      </TouchableOpacity>
-    )}
-    {props.validCertificate.markdownPreview && (
       <>
-        <MarkdownHandleCustomLink testID={"markdownPreview"}>
-          {props.validCertificate.markdownPreview}
-        </MarkdownHandleCustomLink>
+        <View spacer={true} />
+        <TouchableOpacity
+          testID={"QRCode"}
+          accessible={true}
+          accessibilityRole={"imagebutton"}
+          accessibilityLabel={I18n.t(
+            "features.euCovidCertificate.valid.accessibility.qrCode"
+          )}
+          accessibilityHint={I18n.t(
+            "features.euCovidCertificate.valid.accessibility.hint"
+          )}
+          onPress={() =>
+            props.navigateToQrCodeFullScreen(
+              props.validCertificate.qrCode.content
+            )
+          }
+        >
+          <Image
+            source={{
+              uri: `data:image/png;base64,${props.validCertificate.qrCode.content}`
+            }}
+            style={styles.qrCode}
+            onError={() => {
+              void mixpanelTrack("EUCOVIDCERT_QRCODE_IMAGE_NOT_VALID", {
+                messageId: props.messageId
+              });
+            }}
+          />
+        </TouchableOpacity>
         <View spacer={true} />
       </>
+    )}
+    {props.validCertificate.markdownInfo && (
+      <View style={props.markdownWebViewStyle}>
+        <MarkdownHandleCustomLink
+          testID={"markdownPreview"}
+          extraBodyHeight={60}
+        >
+          {props.validCertificate.markdownInfo}
+        </MarkdownHandleCustomLink>
+        <View spacer={true} />
+      </View>
     )}
   </View>
 );
 
-const Footer = (props: Props): React.ReactElement => {
+const showToastError = (error: string = I18n.t("global.genericError")) =>
+  showToast(error);
+const addBottomSheetItem = (config: {
+  title: string;
+  subTitle: string;
+  onPress: () => void;
+}) => (
+  <ButtonDefaultOpacity
+    onPress={config.onPress}
+    style={styles.container}
+    onPressWithGestureHandler={true}
+  >
+    <View style={styles.flexColumn}>
+      <View style={styles.row}>
+        <View style={IOStyles.flex}>
+          <H3 color={"bluegreyDark"} weight={"SemiBold"}>
+            {config.title}
+          </H3>
+          <H5 color={"bluegrey"} weight={"Regular"}>
+            {config.subTitle}
+          </H5>
+        </View>
+        <IconFont name={"io-right"} color={IOColors.blue} size={24} />
+      </View>
+    </View>
+    <View spacer={true} large={true} />
+    <View spacer={true} large={true} />
+  </ButtonDefaultOpacity>
+);
+
+type FooterProps = Props & { onSave: () => void };
+const Footer = (props: FooterProps): React.ReactElement => {
+  const { present: presentBottomSheet, dismiss } = useIOBottomSheet(
+    <View>
+      {addBottomSheetItem({
+        title: I18n.t(
+          "features.euCovidCertificate.save.bottomSheet.saveAsImage.title"
+        ),
+        subTitle: I18n.t(
+          "features.euCovidCertificate.save.bottomSheet.saveAsImage.subTitle"
+        ),
+        onPress: () => {
+          props.onSave();
+          dismiss();
+        }
+      })}
+    </View>,
+    <View style={IOStyles.flex}>
+      <H3 color={"bluegreyDark"} weight={"SemiBold"}>
+        {I18n.t("features.euCovidCertificate.save.bottomSheet.title")}
+      </H3>
+      <H5 color={"bluegrey"} weight={"Regular"}>
+        {I18n.t("features.euCovidCertificate.save.bottomSheet.subTitle")}
+      </H5>
+      <View spacer={true} />
+      <View spacer={true} />
+    </View>,
+    320
+  );
+
   const saveButton = confirmButtonProps(
-    // TODO: add save function with https://pagopa.atlassian.net/browse/IAGP-17
-    constNull,
+    presentBottomSheet,
     I18n.t("global.genericSave")
   );
   const markdownDetails = props.validCertificate.markdownDetails;
@@ -97,13 +215,91 @@ const Footer = (props: Props): React.ReactElement => {
   );
 };
 
-const EuCovidCertValidScreen = (props: Props): React.ReactElement => (
-  <BaseEuCovidCertificateLayout
-    testID={"EuCovidCertValidScreen"}
-    content={<EuCovidCertValidComponent {...props} />}
-    footer={<Footer {...props} />}
-  />
-);
+const screenShotOption: CaptureOptions = {
+  width: Dimensions.get("window").width,
+  format: "jpg",
+  quality: 1.0
+};
+const EuCovidCertValidScreen = (props: Props): React.ReactElement => {
+  const screenShotViewContainer = React.createRef<View>();
+  const [flashAnimationState, setFlashAnimationState] = useState<
+    FlashAnimationState
+  >();
+  const [isCapturingScreenShoot, setIsCapturingScreenShoot] = useState(false);
+  React.useEffect(() => {
+    if (isCapturingScreenShoot) {
+      // at the end of fadeIn animation, the views inside screenShotViewContainerRef
+      // will be captured in an screenshot image
+      setFlashAnimationState("fadeIn");
+    }
+  }, [isCapturingScreenShoot]);
+
+  const saveScreenShoot = () => {
+    // it should not never happen
+    if (screenShotViewContainer.current === null) {
+      showToastError();
+      return;
+    }
+    captureScreenShoot(screenShotViewContainer, screenShotOption, {
+      onSuccess: () =>
+        showToast(I18n.t("features.euCovidCertificate.save.ok"), "success"),
+      onNoPermissions: () =>
+        showToast(I18n.t("features.euCovidCertificate.save.noPermission")),
+      onError: () => showToast(I18n.t("global.genericError")),
+      onEnd: () => {
+        setFlashAnimationState("fadeOut");
+        setIsCapturingScreenShoot(false);
+      }
+    });
+  };
+  return (
+    <BaseEuCovidCertificateLayout
+      testID={"EuCovidCertValidScreen"}
+      content={
+        <View
+          collapsable={false}
+          ref={screenShotViewContainer}
+          style={[IOStyles.flex, { backgroundColor: IOColors.white }]}
+        >
+          {/* add extra space (top,sides,bottom) and padding while capturing the screenshot */}
+          {isCapturingScreenShoot && <View spacer={true} large={true} />}
+          {isCapturingScreenShoot && (
+            <View style={IOStyles.horizontalContentPadding}>
+              <Header />
+            </View>
+          )}
+          {isCapturingScreenShoot && <View spacer={true} large={true} />}
+          <EuCovidCertValidComponent
+            messageId={props.euCovidCert?.messageId}
+            {...props}
+            markdownWebViewStyle={
+              isCapturingScreenShoot
+                ? IOStyles.horizontalContentPadding
+                : undefined
+            }
+          />
+          {isCapturingScreenShoot && <View spacer={true} large={true} />}
+        </View>
+      }
+      footer={
+        <>
+          <Footer
+            {...props}
+            onSave={() => {
+              void mixpanelTrack("EUCOVIDCERT_SAVE_QRCODE");
+              setIsCapturingScreenShoot(true);
+            }}
+          />
+          {/* this view must be the last one, since it must be drawn on top of all */}
+          <FlashAnimatedComponent
+            state={flashAnimationState}
+            onFadeInCompleted={saveScreenShoot}
+          />
+        </>
+      }
+    />
+  );
+};
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
   navigateToQrCodeFullScreen: (qrCodeContent: string) =>
@@ -113,7 +309,9 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
       navigateToEuCovidCertificateMarkdownDetailsScreen({ markdownDetails })
     )
 });
-const mapStateToProps = (_: GlobalState) => ({});
+const mapStateToProps = (state: GlobalState) => ({
+  euCovidCert: euCovidCertCurrentSelector(state)
+});
 
 export default connect(
   mapStateToProps,