+++ ts/store/middlewares/analytics.ts
@@ -1,23 +1,12 @@
 /* eslint-disable no-fallthrough */
 // disabled in order to allows comments between the switch
-import { constNull } from "fp-ts/lib/function";
-import DeviceInfo from "react-native-device-info";
-import { sha256 } from "react-native-sha256";
 import { NavigationActions } from "react-navigation";
 import { getType } from "typesafe-actions";
 import { setInstabugUserAttribute } from "../../boot/configureInstabug";
 import {
-  activateBonusVacanze,
-  cancelBonusVacanzeRequest,
-  checkBonusVacanzeEligibility,
   loadAllBonusActivations,
-  loadAvailableBonuses,
-  storeEligibilityRequestId
+  loadAvailableBonuses
 } from "../../features/bonus/bonusVacanze/store/actions/bonusVacanze";
-import {
-  isActivationResponseTrackable,
-  isEligibilityResponseTrackable
-} from "../../features/bonus/bonusVacanze/utils/bonus";
 import { trackBPayAction } from "../../features/wallet/onboarding/bancomatPay/analytics";
 import { trackCoBadgeAction } from "../../features/wallet/onboarding/cobadge/analytics";
 import { trackPrivativeAction } from "../../features/wallet/onboarding/privative/analytics";
@@ -63,6 +52,7 @@ import {
   removeMessages,
   setMessageReadState
 } from "../actions/messages";
+import { setMixpanelEnabled } from "../actions/mixpanel";
 import {
   updateNotificationInstallationFailure,
   updateNotificationsInstallationToken
@@ -128,6 +118,7 @@ import {
 
 import trackBpdAction from "../../features/bonus/bpd/analytics/index";
 import trackCgnAction from "../../features/bonus/cgn/analytics/index";
+import trackEuCovidCertificateActions from "../../features/euCovidCert/analytics/index";
 import trackBancomatAction from "../../features/wallet/onboarding/bancomat/analytics/index";
 import trackSatispayAction from "../../features/wallet/satispay/analytics/index";
 import {
@@ -164,23 +155,6 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
         SPID_IDP_ID: action.payload.id,
         SPID_IDP_NAME: action.payload.name
       });
-    case getType(profileLoadSuccess):
-      // as soon as we have the user fiscal code, attach the mixpanel
-      // session to the hashed fiscal code of the user
-      const fiscalnumber = action.payload.fiscal_code;
-
-      // Re-identify the user using the hashed fiscal code.
-      // It's important the flow order and the order in which the arguments are passed to the
-      // mp.alias function because the second argument is the 'Main ID' for mixpanel so the events
-      // will be showned in the Main ID page.
-      const identifyAndAlias = sha256(fiscalnumber).then(hash =>
-        mp.identify(hash).then(() => mp.alias(DeviceInfo.getUniqueId(), hash))
-      );
-
-      return Promise.all([
-        mp.track(action.type).then(constNull, constNull),
-        identifyAndAlias.then(constNull, constNull)
-      ]);
 
     case getType(idpLoginUrlChanged):
       return mp.track(action.type, {
@@ -323,8 +297,6 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
     //  Bonus vacanze
     case getType(loadAllBonusActivations.failure):
     case getType(loadAvailableBonuses.failure):
-    case getType(checkBonusVacanzeEligibility.failure):
-    case getType(activateBonusVacanze.failure):
       return mp.track(action.type, {
         reason: action.payload.message
       });
@@ -363,6 +335,7 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
     case getType(updatePin):
     // profile
     case getType(profileUpsert.success):
+    case getType(profileLoadSuccess):
     // userMetadata
     case getType(userMetadataUpsert.request):
     case getType(userMetadataUpsert.success):
@@ -371,7 +344,6 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
     // messages
     case getType(loadMessages.request):
     case getType(loadMessagesCancel):
-    case getType(loadMessage.success):
     // services
     case getType(loadVisibleServices.request):
     case getType(loadVisibleServices.success):
@@ -412,32 +384,15 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
     //  profile First time Login
     case getType(profileFirstLogin):
     // other
+    case getType(loadMessage.success):
     case getType(updateNotificationsInstallationToken):
     // bonus vacanze
     case getType(loadAllBonusActivations.request):
     case getType(loadAllBonusActivations.success):
     case getType(loadAvailableBonuses.success):
     case getType(loadAvailableBonuses.request):
-    case getType(checkBonusVacanzeEligibility.request):
-    case getType(cancelBonusVacanzeRequest):
-    case getType(storeEligibilityRequestId):
       return mp.track(action.type);
 
-    // bonus vacanze
-    case getType(checkBonusVacanzeEligibility.success):
-      if (isEligibilityResponseTrackable(action.payload)) {
-        return mp.track(action.type, {
-          status: action.payload.status
-        });
-      }
-      break;
-    case getType(activateBonusVacanze.success):
-      if (isActivationResponseTrackable(action.payload)) {
-        return mp.track(action.type, {
-          status: action.payload.status
-        });
-      }
-      break;
     case getType(deleteUserDataProcessing.request):
       return mp.track(action.type, { choice: action.payload });
     case getType(removeAccountMotivation):
@@ -447,6 +402,10 @@ const trackAction = (mp: NonNullable<typeof mixpanel>) => (
       return mp.track(action.type, {
         choice: action.payload.choice,
         reason: action.payload.error.message
+      });
+    case getType(setMixpanelEnabled):
+      return mp.track(action.type, {
+        value: action.payload
       });
   }
   return Promise.resolve();
@@ -470,6 +429,7 @@ export const actionTracking = (_: MiddlewareAPI) => (next: Dispatch) => (
     void trackPrivativeAction(mixpanel)(action);
     void trackCgnAction(mixpanel)(action);
     void trackContentAction(mixpanel)(action);
+    void trackEuCovidCertificateActions(mixpanel)(action);
   }
   return next(action);
 };